disallows duplicate imports
I had originally added this because I thought it would help me in the ticket for disallowing unused imports, but it turns out it doesn't do that.  but when I saw what it _does_ do (one of which, is disallowing duplicate imports of which we did have some), it seems like it's work keeping around.