Fix: react-query typescript in connect-query-plugin (#69)
I was discovered that there is an inconsistency with the parameters
passed to the useInfiniteQuery function, particularly when using
use<Name>InfiniteQuery. This inconsistency results in type mismatch
errors.

Current Implementation:

The useInfiniteQuery function has the following signature:

```ts
useInfiniteQuery: <ParamKey extends keyof PartialMessage<I>>(
  input: DisableQuery | PartialMessage<I>,
  options: BaseInfiniteQueryOptions<I, O, ParamKey> & RequireExactlyOne<{
    applyPageParam: (options: {
      pageParam: PartialMessage<I>[ParamKey] | undefined;
      input: PartialMessage<I>;
    }) => PartialMessage<I>;
    pageParamKey: ParamKey;
  }>
) => {
  // Implementation details...
};

```

However, in the useSayInfiniteQuery function, the queryOptions and
options are made optional:
```
export const useSayInfiniteQuery = (
  inputs: Parameters<typeof say.useInfiniteQuery>[0],
  queryOptions?: Partial<UseInfiniteQueryOptions<PartialMessage<SayRequest>, ConnectError>>,
  options?: Parameters<typeof say.useInfiniteQuery>[1],
) => {
  const baseOptions = say.useInfiniteQuery(inputs, options);
  return useInfiniteQuery({
    ...baseOptions,
    ...queryOptions,
  });
};
```
However, the `options` parameter is currently required in
`useInfiniteQuery`. To address this, we propose swapping the order of
queryOptions and options, making `options` required instead. This
adjustment will not only ensure consistency but also enhance code
readability. We'll also need to update the order of parameters for other
hooks to maintain uniformity across the project

**PLEASE HELP HAVE A LOOK**
Also, there is an issue with the useInfiniteQuery function when the
InputRequest is an empty object (PartialMessage<I> = {}). This leads to
a TypeScript error in certain scenarios.

```
useInfiniteQuery: <ParamKey extends keyof PartialMessage<I>>(input: DisableQuery | PartialMessage<I>, options: BaseInfiniteQueryOptions<I, O, ParamKey> & RequireExactlyOne<{
        applyPageParam: (options: {
            pageParam: PartialMessage<I>[ParamKey] | undefined;
            input: PartialMessage<I>;
        }) => PartialMessage<I>;
        pageParamKey: ParamKey;
    }>) => {
        enabled: boolean;
        queryKey: ConnectQueryKey<I>;
        queryFn: (context: QueryFunctionContext<ConnectQueryKey<I>, PartialMessage<I>[ParamKey]>) => Promise<O>;
        getNextPageParam: GetNextPageParamFunction<O>;
        onError?: (error: ConnectError) => void;
    };
```
with
- PartialMessage<I> = {}
- <ParamKey extends keyof PartialMessage<I>> = never

EX in `useWorkInfiniteQuery`
it said 
```
      Types of property 'queryFn' are incompatible.
        Type '(context: QueryFunctionContext<ConnectQueryKey<Nothing>, never>) => Promise<Nothing>' is not assignable to type 'QueryFunction<Nothing, ConnectQueryKey<Nothing>>'.
```
also tried to rteplace ConnectQueryKey = never but it's said 
```
        Type '(context: QueryFunctionContext<ConnectQueryKey<Nothing>, never>) => Promise<Nothing>' is not assignable to type 'QueryFunction<Nothing, ConnectQueryKey<Nothing>>'.

```

![image](https://github.com/bufbuild/connect-query/assets/30227910/b5f02d42-2893-4a80-8307-40dfb00118b5)
I've make an change like this
```
return useInfiniteQuery<Nothing, ConnectError, Nothing, keyof typeof inputs extends never ? any : ConnectQueryKey<Nothing>>({
        ...baseOptions,
        ...queryOptions,
    });
```

if keyof InputRquest is `never`, take `any` to bypass this situation.

Otherwise, it works well in others

![image](https://github.com/bufbuild/connect-query/assets/30227910/3af50023-3f2b-4074-b24d-51d4ed17d759)