Release 0.4.2 (#119)
## What's Changed

To keep Connect well-maintained and responsive to its users' needs over
the long term, we're preparing to donate it to a foundation. (More
details on that soon!) To cleanly separate Connect from Buf's other
code, we're moving development to the
[connectrpc](https://github.com/connectrpc) GitHub organization and to
the [connectrpc](https://www.npmjs.com/org/connectrpc) organization on
npmjs.com.

This is the first release that publishes packages with the new
`@connectrpc` scope. To make the switch seamless, we are introducing a
small tool that updates all references in your project automatically,
[@connectrpc/connect-migrate](https://www.npmjs.com/package/@connectrpc/connect-migrate).

The switch is as simple as running a single command:

```
$ npm install @connectrpc/connect-migrate
$ npx connect-migrate
Scanning... ✓
    1 package.json file
    1 lock file
    5 source files
Updating source files... 
  src/client.ts ✓
  src/server.ts ✓
  src/webclient.ts ✓
Updating packages... ✓
  package.json ✓
Updating lock file... 
  package-lock.json ✓
```

Before you migrate, make sure that you are on the latest release of the
`@bufbuild` packages related to Connect.

| Old package | New package |

|-------------------------------------------------|---------------------------------------------------|
| @bufbuild/connect v0.13.0 | @connectrpc/connect v0.13.1 |
| @bufbuild/connect-web v0.13.0 | @connectrpc/connect-web v0.13.1 |
| @bufbuild/connect-fastify v0.13.0 | @connectrpc/connect-fastify
v0.13.1 |
| @bufbuild/connect-node v0.13.0 | @connectrpc/connect-node v0.13.1 |
| @bufbuild/connect-next v0.13.0 | @connectrpc/connect-next v0.13.1 |
| @bufbuild/connect-express v0.13.0 | @connectrpc/connect-express
v0.13.1 |
| @bufbuild/protoc-gen-connect-es v0.13.0 |
@connectrpc/protoc-gen-connect-es v0.13.1 |
| @bufbuild/protoc-gen-connect-web v0.13.0 |
@connectrpc/protoc-gen-connect-web v0.13.1 |
| @bufbuild/connect-query v0.4.1 | @connectrpc/connect-query v0.4.2 |
| @bufbuild/protoc-gen-connect-query v0.4.1 |
@connectrpc/protoc-gen-connect-query v0.4.2 |
| @bufbuild/protoc-gen-connect-query-react v0.4.1 |
@connectrpc/protoc-gen-connect-query-react v0.4.2 |


**Full Changelog**:
https://github.com/connectrpc/connect-query-es/compare/v0.4.1...v0.4.2