Extract React specific dependencies into single file (#169)
Less reliance on React specific APIs (in the core of connect-query)
allows the core API to adapt better to scenarios like Server Components
or other frameworks.

This PR keeps the existing plugin API (while adding some new methods)
but frees us up to be more selective about the react APIs we NEED for
other future plugins.

## Changes in `@connectrpc/connect-query`

### Breaking changes
- Removed `UnaryHooks` type and replaced with `UnaryFunctionsWithHooks`
- Updated `createUnaryHooks` to only return the hooks and not other
functions
 
### New functions

### createUnaryFunctions

Introduced `createUnaryFunctions` which can be used with the new
`createUnaryHooks` to do the same thing as the old `createUnaryHooks`.
The reasoning behind this change is to prep the core library,
`connect-query`, to be able to used separately from React itself and
this is just the first step in that direction.

## Changes in code generated by `protoc-gen-connect-query` and
`protoc-gen-connect-query-react`

### New functions

Each service method now generates two new functions associated to the
method:

#### method.createUseInfiniteQueryOptions

This method creates options that will pass to `useInfiniteQuery`. It is
identical to `method.useInfiniteQuery` except it needs to be passed a
transport directly.

#### createUseMutationOptions

This method creates options that will pass to `useMutation`. It is
identical to `method.useMutation` except it needs to be passed a
transport directly.

We added these methods to better future-proof the code for incoming
changes to React (React Server Components, etc).