Remove experimental plugin (#178)
`@tanstack/react-query` v5 is introducing some major changes to their
types and also introduce suspense specific APIs (useSuspenseQuery,
useSuspenseInfiniteQuery, and useSuspenseQueries). These APIs are pretty
easy to integrate with the original plugin (non-react specific) but
would need an additional two APIs generated for every service method.
This starts to create a lot of noise with our auto-complete options:

As an example, if we have a service method called `say`, we now must
generate `useSayQuery`, `useSayInfiniteQuery`, `useSayMutation`,
`useSaySuspenseQuery`, and `useSaySuspenseInfiniteQuery`. This is
increasingly making a lot of noise when just trying to use the `say`
method from a service. This is much worse with larger service method
names.

In keeping with best practices (limiting the number of ways to do a
single thing), we've decided to remove this plugin from the repo.