Fix missing implicit types (#211)
We have really appreciated this lib, and have been using it since it
came out! I think I found a typing fix to make it even more usable.

When trying to bundle up the generated code with `tsc
--declaration=true`, ts couldn't figure out the `disableQuery` type and
the `BaseInfiniteQueryOptions` type. This is able to be resolved if the
exported queries are explicitly typed.

This also will resolve #209 🎉 

Let me know if there is anything else I can do to this PR!