Migrate to Node16 resolution (#186)
Updating everything to use node16 resolutions. This guarantees (via the
tsc compiler) that the files we output match the expected import format
for downstream tools. This also brings us into compliance with module
resolution for node 16+, which is the minimum version of node we support
and is the recommended value for libraries.

See detailed explanation
[here](https://www.typescriptlang.org/docs/handbook/esm-node.html)