Upgrade vitest to 3.2.4 (#551)
This upgrades to Vitest v3.2.4 to fix a critical security issue on a
transitive dependency.

The issue is
[here](https://github.com/connectrpc/connect-query-es/security/dependabot/30)
and occurs with `form-data` 4.0.1. which we pull in as follows:
```
└─┬ vitest@3.1.4
  └─┬ jsdom@20.0.3
    └── form-data@4.0.1
```

Upgrading to vitest 3.2.4 kicks the package-lock dep tree to now be:

```
└─┬ vitest@3.2.4
  └─┬ jsdom@20.0.3
    └── form-data@4.0.4
```

As part of these updates, we are regenerating the package-lock which
also causes two additional changes:

* New generated files for the upgrade of protobuf-es.
* There seems to be a breaking change in react-query and
`InfiniteQueryOptions` (see commit
[here](https://github.com/TanStack/query/commit/efa963d1fd90d408d3500ef8754c78fb243b9473#diff-339c38a0d756225f7e7aad7232ec545d86b4863e5d474101b1a9cf1e1c1fb898))
which is tickled by this upgrade. The additional fixes to
`use-infinite-query.ts` are a result of this.

---------

Signed-off-by: Steve Ayers <sayers@buf.build>