Ignore page param when generating infinite key in createInfiniteQueryOptions (#559)
Depending on the incoming input, infinite query keys should ignore
whatever property informs the page to be returned. This was being done
properly when calling `callUnaryMathod` but was not being propagated to
`createConnectQueryKey`. This means that calls to `getQueryData` or
`setQueryData` would not find the data initialized by
`useInfiniteQuery`.

The original test failed to catch this case since 0n (and 0) are default
values that evaluate to the same key. Adding a test with non-default
page param identified the issue.

---------

Signed-off-by: Paul Sachs <psachs@buf.build>
Signed-off-by: Paul Sachs <11449728+paul-sachs@users.noreply.github.com>